CLD3::NNetLanguageIdentifier#find_language may return nil
CLD3::NNetLanguageIdentifier#find_language used to return a Result whose
language is nil when it cannot determine the language. However, the other
attributes of the returned object is useless in the case. There is no
obstacle to return nil instead of returning such an object for Ruby in
contrary to the C++ function constrained by type.