Rewrite as an extension
The initial intention of using FFI was to make it compatible with
Rubinius and JRuby just in case we need to use these platforms.

However, Rubinius is no longer maintained, and TruffleRuby is becoming a
promising alternative to JRuby, and it supports extensions. It's time to
ditch FFI.

This fixes the incompatibility with the latest RubyGems which was
discussed at:
https://github.com/rubygems/rubygems/pull/6218